# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Game Dev IDE API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export interface Project {
  id: string;
  name: string;
  description?: string;
  createdAt: string;
  updatedAt: string;
  entryFile?: string;
  fileCount: number;
}

export interface CreateProjectBody {
  name: string;
  description?: string;
}

export interface UpdateProjectBody {
  name?: string;
  description?: string;
}

export type FileNodeType = (typeof FileNodeType)[keyof typeof FileNodeType];

export const FileNodeType = {
  file: "file",
  folder: "folder",
} as const;

export interface FileNode {
  name: string;
  path: string;
  type: FileNodeType;
  children?: FileNode[];
  size?: number;
}

export interface FileContent {
  path: string;
  content: string;
  language?: string;
}

export interface WriteFileBody {
  path: string;
  content: string;
}

export interface CreateFolderBody {
  path: string;
}

export interface RenameFileBody {
  oldPath: string;
  newName: string;
}

export interface MoveFileBody {
  sourcePath: string;
  destinationFolder: string;
}

export interface DuplicateFileBody {
  path: string;
}

export interface UploadResult {
  uploaded: string[];
  skipped: string[];
  errors: string[];
}

export interface AiChatBody {
  message: string;
  contextFiles?: string[];
}

export type AiMessageRole = (typeof AiMessageRole)[keyof typeof AiMessageRole];

export const AiMessageRole = {
  user: "user",
  assistant: "assistant",
} as const;

export interface AiMessage {
  id: string;
  role: AiMessageRole;
  content: string;
  timestamp: string;
  thinking?: string;
  changesCount?: number;
  thinkingDurationMs?: number;
}

export interface FileChange {
  id: string;
  projectId: string;
  filePath: string;
  previousContent?: string;
  newContent: string;
  timestamp: string;
  aiMessageId?: string;
  description?: string;
}

export interface PreviewEntry {
  entryFile: string;
  url: string;
  generated?: boolean;
}

export type ReadFileParams = {
  path: string;
};

export type DeleteFileParams = {
  path: string;
};

export type UploadFilesBody = {
  files?: Blob[];
  paths?: string[];
  overwrite?: boolean;
};

export type UploadZipBody = {
  file?: Blob;
  overwrite?: boolean;
};