# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: AI Game Dev IDE API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects/{id}/files/read:
    get:
      operationId: readFile
      tags: [files]
      summary: Read file content
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: path
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: File content
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileContent"
  /projects/{id}/files/write:
    post:
      operationId: writeFile
      tags: [files]
      summary: Write file content
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WriteFileBody"
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
...
    FileContent:
      type: object
      properties:
        path:
          type: string
        content:
          type: string
        language:
          type: string
      required:
        - path
        - content
    WriteFileBody:
      type: object
      properties:
        path:
          type: string
        content:
          type: string
      required:
        - path
        - content
...