# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Game Dev IDE API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all projects
 */
export const ListProjectsResponseItem = zod.object({
  id: zod.string(),
  name: zod.string(),
  description: zod.string().optional(),
  createdAt: zod.date(),
  updatedAt: zod.date(),
  entryFile: zod.string().optional(),
  fileCount: zod.number(),
});
export const ListProjectsResponse = zod.array(ListProjectsResponseItem);

/**
 * @summary Create a new project
 */
export const CreateProjectBody = zod.object({
  name: zod.string(),
  description: zod.string().optional(),
});

/**
 * @summary Get a project
 */
export const GetProjectParams = zod.object({
  id: zod.coerce.string(),
});

export const GetProjectResponse = zod.object({
  id: zod.string(),
  name: zod.string(),
  description: zod.string().optional(),
  createdAt: zod.date(),
  updatedAt: zod.date(),
  entryFile: zod.string().optional(),
  fileCount: zod.number(),
});

/**
 * @summary Update a project
 */
export const UpdateProjectParams = zod.object({
  id: zod.coerce.string(),
});

export const UpdateProjectBody = zod.object({
  name: zod.string().optional(),
  description: zod.string().optional(),
});

export const UpdateProjectResponse = zod.object({
  id: zod.string(),
  name: zod.string(),
  description: zod.string().optional(),
  createdAt: zod.date(),
  updatedAt: zod.date(),
  entryFile: zod.string().optional(),
  fileCount: zod.number(),
});

/**
 * @summary Delete a project
 */
export const DeleteProjectParams = zod.object({
  id: zod.coerce.string(),
});

/**
 * @summary Duplicate a project
 */
export const DuplicateProjectParams = zod.object({
  id: zod.coerce.string(),
});

/**
 * @summary List project files as tree
 */
export const ListFilesParams = zod.object({
  id: zod.coerce.string(),
});

export const ListFilesResponseItem = zod.object({
  name: zod.string(),
  path: zod.string(),
  type: zod.enum(["file", "folder"]),
  children: zod.array(zod.unknown()).optional(),
  size: zod.number().optional(),
});
export const ListFilesResponse = zod.array(ListFilesResponseItem);

/**
 * @summary Read file content
 */
export const ReadFileParams = zod.object({
  id: zod.coerce.string(),
});

export const ReadFileQueryParams = zod.object({
  path: zod.coerce.string(),
});

export const ReadFileResponse = zod.object({
  path: zod.string(),
  content: zod.string(),
  language: zod.string().optional(),
});

/**
 * @summary Write file content
 */
export const WriteFileParams = zod.object({
  id: zod.coerce.string(),
});

export const WriteFileBody = zod.object({
  path: zod.string(),
  content: zod.string(),
});

export const WriteFileResponse = zod.object({
  success: zod.boolean(),
  message: zod.string().optional(),
});

/**
 * @summary Create a folder
 */
export const CreateFolderParams = zod.object({
  id: zod.coerce.string(),
});

export const CreateFolderBody = zod.object({
  path: zod.string(),
});

export const CreateFolderResponse = zod.object({
  success: zod.boolean(),
  message: zod.string().optional(),
});

/**
 * @summary Rename a file or folder
 */
export const RenameFileParams = zod.object({
  id: zod.coerce.string(),
});

export const RenameFileBody = zod.object({
  oldPath: zod.string(),
  newName: zod.string(),
});

export const RenameFileResponse = zod.object({
  success: zod.boolean(),
  message: zod.string().optional(),
});

/**
 * @summary Move a file or folder
 */
export const MoveFileParams = zod.object({
  id: zod.coerce.string(),
});

export const MoveFileBody = zod.object({
  sourcePath: zod.string(),
  destinationFolder: zod.string(),
});

export const MoveFileResponse = zod.object({
  success: zod.boolean(),
  message: zod.string().optional(),
});

/**
 * @summary Delete a file or folder
 */
export const DeleteFileParams = zod.object({
  id: zod.coerce.string(),
});

export const DeleteFileQueryParams = zod.object({
  path: zod.coerce.string(),
});

export const DeleteFileResponse = zod.object({
  success: zod.boolean(),
  message: zod.string().optional(),
});

/**
 * @summary Duplicate a file
 */
export const DuplicateFileParams = zod.object({
  id: zod.coerce.string(),
});

export const DuplicateFileBody = zod.object({
  path: zod.string(),
});

export const DuplicateFileResponse = zod.object({
  success: zod.boolean(),
  message: zod.string().optional(),
});

/**
 * @summary Upload files (multipart)
 */
export const UploadFilesParams = zod.object({
  id: zod.coerce.string(),
});

export const UploadFilesBody = zod.object({
  files: zod.array(zod.instanceof(File)).optional(),
  paths: zod.array(zod.string()).optional(),
  overwrite: zod.boolean().optional(),
});

export const UploadFilesResponse = zod.object({
  uploaded: zod.array(zod.string()),
  skipped: zod.array(zod.string()),
  errors: zod.array(zod.string()),
});

/**
 * @summary Upload a zip file and extract it
 */
export const UploadZipParams = zod.object({
  id: zod.coerce.string(),
});

export const UploadZipBody = zod.object({
  file: zod.instanceof(File).optional(),
  overwrite: zod.boolean().optional(),
});

export const UploadZipResponse = zod.object({
  uploaded: zod.array(zod.string()),
  skipped: zod.array(zod.string()),
  errors: zod.array(zod.string()),
});

/**
 * @summary Send a message to the AI coding agent (SSE streaming)
 */
export const AiChatParams = zod.object({
  id: zod.coerce.string(),
});

export const AiChatBody = zod.object({
  message: zod.string(),
  contextFiles: zod.array(zod.string()).optional(),
});

/**
 * @summary Get AI conversation history
 */
export const GetAiHistoryParams = zod.object({
  id: zod.coerce.string(),
});

export const GetAiHistoryResponseItem = zod.object({
  id: zod.string(),
  role: zod.enum(["user", "assistant"]),
  content: zod.string(),
  timestamp: zod.date(),
  thinking: zod.string().optional(),
  changesCount: zod.number().optional(),
  thinkingDurationMs: zod.number().optional(),
});
export const GetAiHistoryResponse = zod.array(GetAiHistoryResponseItem);

/**
 * @summary Clear AI conversation history
 */
export const ClearAiHistoryParams = zod.object({
  id: zod.coerce.string(),
});

/**
 * @summary Get file change history
 */
export const GetChangeHistoryParams = zod.object({
  id: zod.coerce.string(),
});

export const GetChangeHistoryResponseItem = zod.object({
  id: zod.string(),
  projectId: zod.string(),
  filePath: zod.string(),
  previousContent: zod.string().optional(),
  newContent: zod.string(),
  timestamp: zod.date(),
  aiMessageId: zod.string().optional(),
  description: zod.string().optional(),
});
export const GetChangeHistoryResponse = zod.array(GetChangeHistoryResponseItem);

/**
 * @summary Revert a file to a previous version
 */
export const RevertChangeParams = zod.object({
  id: zod.coerce.string(),
  changeId: zod.coerce.string(),
});

export const RevertChangeResponse = zod.object({
  success: zod.boolean(),
  message: zod.string().optional(),
});

/**
 * @summary Get project preview entry point
 */
export const GetPreviewEntryParams = zod.object({
  id: zod.coerce.string(),
});

export const GetPreviewEntryResponse = zod.object({
  entryFile: zod.string(),
  url: zod.string(),
  generated: zod.boolean().optional(),
});